fixes snap threshold
This does not affect circular nodes, but it does affect nodes where he width and height differ by a lot. The edge should snap when the cursor is at a certain distance from the border of the element, not when it is at a certain distance from the center of the element.